requirements.txt
# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Werkzeug==3.0.1

# Database
alembic==1.13.1
Flask-Migrate==4.0.5
psycopg[binary]==3.2.6

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Caching & Queue
redis==5.0.1
Flask-Caching==2.1.0
celery==5.3.4

# Security
Flask-Talisman==1.1.0
bleach==6.1.0

# Utilities
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.1.2
g4f==0.5.7.5  # MikuGPT AI - DO NOT CHANGE
langdetect==1.0.9
captcha==0.5.0
Pillow==11.3.0
APScheduler==3.10.4
cloudinary==1.41.0

# Production
gunicorn==21.2.0
waitress==2.1.2
